pyproject.toml
[project]
name = "deepseek-engineer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.58.1",
    "prompt-toolkit>=3.0.50",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "fastapi>=0.110.0",
    "sse-starlette>=1.6.5",
]
